backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
psycopg2-binary
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
coverage>=7.0
langchain
langchain-core
langchain-google-genai
langgraph
pydantic
typing-extensions
django-ratelimit==4.1.0
drf-spectacular==0.27.0
